Fix for GPU buffer corruption bug causing GPU crashes and hangs.
The problem was a cache issue.
The soultion currently adopted is to use the DMA API to allocate the
buffers.

This works but is non optimal in terms of memory fragmentation since
the allocated buffers are now physically contiguous although this is
not required due to the availability of the GPU's MMU.

See comments in gsl_linux_map.c for alternative strategies.